Add:submit unsubmitted code

+++ 8.1Linear_time_sorting/BucketSort.go
@@ -1,7 +1,7 @@
 package Linear_time_sorting
 
 type Node struct {
-	value float64
+	value int
 	Next  *Node
 }
